Update version to 1.0.0
I believe the plan was to refactor the Urlbox::Client to have methods
like: #create_render, rather then #get, but as this hasn't happened, I
think it makes sense to increase the verison number to 1.0.0. So that
this is now a non-beta release.

Any big change such as using methods, called #create_render, etc, rather
than wrapping/ mapping to the underlying get request to the UrlBox API, can
be done as a version 2? It will need a fair bit of documentation though
as using #get, #post, etc as this does, maps quickly nicely to the existing
API docs.